nova/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mysqlclient==2.2.4
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
